Poistin Size-metodista varoituksen static_cast:illa.

+++ include/NFmiPath.h
@@ -75,7 +75,7 @@ namespace Imagine
 	
 	const NFmiPathData & Elements(void) const { return itsElements; }
 	
-	int Size(void) const { return itsElements.size(); }
+	int Size(void) const { return static_cast<int>(itsElements.size()); }
 	int Empty(void) const { return itsElements.empty(); }
 	
 	// Clear contents
